Extract selfhost org sub page
