+++ apps/web/src/app/organizations/billing/organization-billing.module.ts
@@ -9,6 +9,7 @@ import { DownloadLicenseComponent } from "./download-license.component";
 import { OrgBillingHistoryViewComponent } from "./organization-billing-history-view.component";
 import { OrganizationBillingRoutingModule } from "./organization-billing-routing.module";
 import { OrganizationBillingTabComponent } from "./organization-billing-tab.component";
+import { OrganizationSubscriptionSelfhostComponent } from "./organization-subscription-selfhost.component";
 import { OrganizationSubscriptionComponent } from "./organization-subscription.component";
 
 @NgModule({
@@ -21,6 +22,7 @@ import { OrganizationSubscriptionComponent } from "./organization-subscription.c
     OrganizationBillingTabComponent,
     OrganizationSubscriptionComponent,
     OrgBillingHistoryViewComponent,
+    OrganizationSubscriptionSelfhostComponent,
   ],
 })
 export class OrganizationBillingModule {}
