+++ apps/web/src/app/organizations/billing/organization-subscription.component.html
@@ -238,76 +238,12 @@
       </button>
     </div>
   </ng-container>
-  <ng-container *ngIf="selfHosted">
-    <dl>
-      <dt>{{ "billingPlan" | i18n }}</dt>
-      <dd>{{ sub.plan.name }}</dd>
-      <dt>{{ "expiration" | i18n }}</dt>
-      <dd *ngIf="sub.expiration">
-        {{ sub.expiration | date: "mediumDate" }}
-        <span *ngIf="isExpired" class="text-danger ml-2">
-          <i class="bwi bwi-exclamation-triangle" aria-hidden="true"></i>
-          {{ "licenseIsExpired" | i18n }}
-        </span>
-      </dd>
-      <dd *ngIf="!sub.expiration">{{ "neverExpires" | i18n }}</dd>
-    </dl>
-    <div>
-      <button type="button" class="btn btn-outline-secondary" (click)="updateLicense()">
-        {{ "updateLicense" | i18n }}
-      </button>
-      <a
-        href="https://vault.bitwarden.com"
-        target="_blank"
-        rel="noopener"
-        class="btn btn-outline-secondary"
-      >
-        {{ "manageSubscription" | i18n }}
-      </a>
-    </div>
-    <div class="card mt-3" *ngIf="showUpdateLicense">
-      <div class="card-body">
-        <button
-          type="button"
-          class="close"
-          appA11yTitle="{{ 'cancel' | i18n }}"
-          (click)="closeUpdateLicense(false)"
-        >
-          <span aria-hidden="true">&times;</span>
-        </button>
-        <h3 class="card-body-header">{{ "updateLicense" | i18n }}</h3>
-        <app-update-license
-          [organizationId]="organizationId"
-          (onUpdated)="closeUpdateLicense(true)"
-          (onCanceled)="closeUpdateLicense(false)"
-        ></app-update-license>
-      </div>
-    </div>
-    <div *ngIf="showBillingSyncKey">
-      <h2 class="mt-5">
-        {{ "billingSync" | i18n }}
-      </h2>
-      <p>
-        {{ "billingSyncDesc" | i18n }}
-      </p>
-      <button
-        type="button"
-        class="btn btn-outline-secondary"
-        (click)="manageBillingSyncSelfHosted()"
-      >
-        {{ "manageBillingSync" | i18n }}
-      </button>
-      <small class="form-text text-muted" *ngIf="billingSyncSetUp">
-        {{ "lastSync" | i18n }}:
-        <span *ngIf="userOrg.familySponsorshipLastSyncDate != null">
-          {{ userOrg.familySponsorshipLastSyncDate | date: "medium" }}
-        </span>
-        <span *ngIf="userOrg.familySponsorshipLastSyncDate == null">
-          {{ "never" | i18n | lowercase }}
-        </span>
-      </small>
-    </div>
-  </ng-container>
+  <app-org-subscription-selfhost
+    *ngIf="selfHosted"
+    [sub]="sub"
+    [organizationId]="organizationId"
+    [userOrg]="userOrg"
+    (reload)="load()"
+  ></app-org-subscription-selfhost>
 </ng-container>
 <ng-template #setupBillingSyncTemplate></ng-template>
-<ng-template #rotateBillingSyncKeyTemplate></ng-template>
